fix: 模组金额类型统一为 long，信用额度基于终身总收入减贷款本金
- BankData 中 CheckingBalance, InterestEarned, FixedDeposit.Amount,
  LoanAccount.Principal, LoanAccount.InterestAccrued 改为 long
- BankCalculator 返回类型相应改为 long
- Farmer.Money 交互处加 (int) 强转（SDV 原版为 int）
- 信用额度基数 = totalMoneyEarned - 历史上所有贷款本金之和

+++ Red Panda Bazaar Code/Features/Bank/BankCheckingMenu.cs
@@ -93,8 +93,8 @@ public class BankCheckingMenu : IClickableMenu
                     if (Context.IsMainPlayer && number > 0)
                         Game1.activeClickableMenu = new BankCheckingMenu();
                 },
-                price: -1, minValue: 1, maxValue: Math.Max(1, Bank.GetCheckingBalance()),
-                defaultNumber: Math.Min(100, Bank.GetCheckingBalance()));
+                price: -1, minValue: 1, maxValue: (int)Math.Max(1L, Bank.GetCheckingBalance()),
+                defaultNumber: (int)Math.Min(100L, Bank.GetCheckingBalance()));
         }
     }
